[JBTM-2633] Update arquillian.xml in tests to contain settings for server startup timeout
In case that test is run against some slow database on a slow machine
when jdbc  object store is used then server could not be started on time
of 60 seconds and test start to fail.
This change adds configurable option to increase the arquillian settings
of server startup timeout if necessary.